+++ src/main/java/org/elasticsearch/search/reducers/metric/single/sum/Sum.java
@@ -17,20 +17,11 @@
  * under the License.
  */
 
-package org.elasticsearch.search.reducers.metric.sum;
+package org.elasticsearch.search.reducers.metric.single.sum;
 
-import org.elasticsearch.search.aggregations.Aggregation;
-import org.elasticsearch.search.aggregations.Aggregations;
-import org.elasticsearch.search.aggregations.InternalAggregation;
-import org.elasticsearch.search.reducers.Reducer;
-import org.elasticsearch.search.reducers.ReducerContext;
-import org.elasticsearch.search.reducers.ReducerFactories;
 import org.elasticsearch.search.reducers.ReductionExecutionException;
 import org.elasticsearch.search.reducers.metric.MetricOp;
-import org.elasticsearch.search.reducers.metric.SimpleMetricReducer;
-import org.elasticsearch.search.reducers.metric.SingleMetricResult;
-
-import java.io.IOException;
+import org.elasticsearch.search.reducers.metric.single.SingleMetricResult;
 
 public class Sum implements MetricOp {
